backend/pyproject.toml
[project]
name = "cohab-backend"
version = "1.0.0"
description = "Backend API for Cohab roommate management app"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strawberry-graphql[fastapi]>=0.219.1",
    "fastapi==0.104.1",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "supabase>=2.23.0",
]
